Revert "Merge branch 'develop' into fix-layer-assignemnts"
This reverts commit bc97df6a18794bba0d8c70e56b3456fbe25cc9c1.

+++ gateway/src/commands/run.rs
@@ -1,13 +1,14 @@
 // Copyright 2020-2023 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::support::config::build_config;
 use crate::{
-    commands::{ensure_config_version_compatibility, OverrideConfig},
+    commands::{override_config, version_check, OverrideConfig},
+    config::Config,
     OutputFormat,
 };
 use clap::Args;
-use std::error::Error;
+use config::NymConfig;
+use log::*;
 use std::net::IpAddr;
 use std::path::PathBuf;
 use validator_client::nyxd;
@@ -49,12 +50,12 @@ pub struct Run {
     nym_apis: Option<Vec<url::Url>>,
 
     /// Comma separated list of endpoints of the validator
+    #[cfg(feature = "coconut")]
     #[clap(
         long,
         alias = "validators",
-        alias = "nyxd_validators",
-        value_delimiter = ',',
-        hide = true
+        alias = "nymd_validators",
+        value_delimiter = ','
     )]
     // the alias here is included for backwards compatibility (1.1.4 and before)
     nyxd_urls: Option<Vec<url::Url>>,
@@ -65,7 +66,8 @@ pub struct Run {
 
     /// Set this gateway to work only with coconut credentials; that would disallow clients to
     /// bypass bandwidth credential requirement
-    #[clap(long, hide = true)]
+    #[cfg(feature = "coconut")]
+    #[clap(long)]
     only_coconut_credentials: Option<bool>,
 
     /// Enable/disable gateway anonymized statistics that get sent to a statistics aggregator server
@@ -91,43 +93,61 @@ impl From<Run> for OverrideConfig {
 
             enabled_statistics: run_config.enabled_statistics,
             statistics_service_url: run_config.statistics_service_url,
+
+            #[cfg(feature = "coconut")]
             nyxd_urls: run_config.nyxd_urls,
+            #[cfg(feature = "coconut")]
             only_coconut_credentials: run_config.only_coconut_credentials,
         }
     }
 }
 
 fn show_binding_warning(address: String) {
-    eprintln!("\n##### NOTE #####");
-    eprintln!(
+    println!("\n##### NOTE #####");
+    println!(
         "\nYou are trying to bind to {} - you might not be accessible to other nodes\n\
          You can ignore this warning if you're running setup on a local network \n\
          or have set a custom 'announce-host'",
         address
     );
-    eprintln!("\n\n");
+    println!("\n\n");
 }
 
 fn special_addresses() -> Vec<&'static str> {
     vec!["localhost", "127.0.0.1", "0.0.0.0", "::1", "[::1]"]
 }
 
-pub async fn execute(args: Run, output: OutputFormat) -> Result<(), Box<dyn Error + Send + Sync>> {
-    let id = args.id.clone();
-    println!("Starting gateway {id}...");
+pub async fn execute(args: &Run, output: OutputFormat) {
+    println!("Starting gateway {}...", args.id);
 
-    let config = build_config(id, args)?;
-    ensure_config_version_compatibility(&config)?;
+    let mut config = match Config::load_from_file(Some(&args.id)) {
+        Ok(cfg) => cfg,
+        Err(err) => {
+            error!(
+                "Failed to load config for {}. Are you sure you have run `init` before? (Error was: {})",
+                args.id,
+                err,
+            );
+            return;
+        }
+    };
+
+    config = override_config(config, OverrideConfig::from(args.clone()));
+
+    if !version_check(&config) {
+        error!("failed the local version check");
+        return;
+    }
 
     if special_addresses().contains(&&*config.get_listening_address().to_string()) {
         show_binding_warning(config.get_listening_address().to_string());
     }
 
     let mut gateway = crate::node::create_gateway(config).await;
-    eprintln!(
+    println!(
         "\nTo bond your gateway you will need to install the Nym wallet, go to https://nymtech.net/get-involved and select the Download button.\n\
          Select the correct version and install it to your machine. You will need to provide the following: \n ");
-    gateway.print_node_details(output)?;
+    gateway.print_node_details(output);
 
-    gateway.run().await
+    gateway.run().await;
 }
